+++ src/versions/v2/models/update-installment-response.ts
@@ -15,26 +15,23 @@
 
 // May contain unused imports in some cases
 // @ts-ignore
-import { ActivityItem } from './activity-item';
-// May contain unused imports in some cases
-// @ts-ignore
-import { GetActivitiesResponseAllOfAdditionalData } from './get-activities-response-all-of-additional-data';
+import { GetInstallmentsResponseDataInner } from './get-installments-response-data-inner';
 
 /**
 * 
 * @export
-* @interface GetActivitiesResponseAllOf
+* @interface UpdateInstallmentResponse
 */
-export interface GetActivitiesResponseAllOf {
+export interface UpdateInstallmentResponse {
     /**
-    * Activities array
-    * @type {Array<ActivityItem>}
+    * If the response is successful or not
+    * @type {boolean}
     */
-    'data'?: Array<ActivityItem>;
+    'success'?: boolean;
     /**
     * 
-    * @type {GetActivitiesResponseAllOfAdditionalData}
+    * @type {GetInstallmentsResponseDataInner}
     */
-    'additional_data'?: GetActivitiesResponseAllOfAdditionalData;
+    'data'?: GetInstallmentsResponseDataInner;
 }